event detection project/requirements.txt
# pip install -r requirements.txt

numpy
Pillow
torch
torchvision
